requirements.txt
pytest==5.1.1
google-cloud-bigquery==1.18.0
Jinja2==2.10.1
PyYAML==5.1.2